начал писать нейронную сеть с нуля
психанул и удалил все к хуям

+++ ChatBot/ChatBot/Test.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Tensorflow;
 
 namespace ChatBot
 {
@@ -12,19 +11,19 @@ namespace ChatBot
         public Test()
         {
             NeuralNetwork network = new NeuralNetwork(2, 1, 2);
-            network.study(new int[,] {
+            network.study(new double[,] {
                 { 1, 1},
             }, 1);
-            network.study(new int[,] {
+            network.study(new double[,] {
                 { 0, 1},
             }, 1);
-            network.study(new int[,] {
+            network.study(new double[,] {
                 { 1, 0},
             }, 1);
-            network.study(new int[,] {
+            network.study(new double[,] {
                 { 0, 0},
             }, 0);
-            Console.WriteLine(network.getAnswer(new int[,] {
+            Console.WriteLine(network.getAnswer(new double[,] {
                 { 0, 0},
             }));
         }
